requirements.txt
requests>=2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pyyaml>=6.0.1
pydantic>=2.6.0
